Simplify Central SSO by eliminating the external token end-point https://issues.apache.org/jira/browse/USERGRID-1267
