+++ stack/services/src/main/java/org/apache/usergrid/management/ManagementService.java
@@ -127,9 +127,6 @@ public interface ManagementService {
 
     public String getActivationTokenForOrganization( UUID organizationId, long ttl ) throws Exception;
 
-    /** Import an Admin User token generated by some other system */
-    public void importTokenForAdminUser( UUID userId, String token, long ttl ) throws Exception;
-
     public ServiceResults getAdminUserActivities( UserInfo user ) throws Exception;
 
     public ServiceResults getAdminUserActivity( UserInfo user ) throws Exception;
